pyproject.toml
[tool.poetry]
name = "dialogy"
version = "0.9.4"
description = "Dialogy is a library for building and managing SLU applications."
authors = ["Amresh Venugopal <amresh.venugopal@gmail.com>"]
license = "MIT"
repository = "https://github.com/skit-ai/dialogy"

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
attrs = "^20.3.0"
requests = "^2.25.1"
pytz = "^2020.4"
python-dateutil = "^2.8.1"
pydash = "^4.9.3"
numpy = "^1.20.2"
scipy = "^1.7.1"
loguru = "^0.5.3"
pandas = "^1.3.2"
sklearn = "^0.0"
tqdm = "^4.62.2"
jiwer = "^2.2.0"
stanza = "^1.3.0"
thefuzz = "^0.19.0"
xgboost = "^1.5.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.0"
pytest-cov = "^2.10.1"
copier = "^5.1.0"
mypy = "^0.790"
httpretty = "^1.0.3"
black = "^20.8b1"
rope = "^0.18.0"
bandit = "^1.7.0"
pylint = "^2.6.0"
watchdog = "^1.0.2"
isort = "^5.7.0"
furo = "^2021.4.11-beta.34"
toml = "^0.10.2"
Sphinx = "4.4.0"
sphinx-copybutton = "^0.3.1"
sphinx-autodoc-typehints = "^1.16.0"
sphinxcontrib-mermaid = "^0.7.1"

[tool.poetry.scripts]
dialogy = "dialogy.cli:main"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
